:bug: fix : 카드 드랍 시 이전 위치로 돌아가는 현상 해결
